+++ Millenium/Controllers/CollectionViewController/CollectionListViewController/CollectionListViewController.m
@@ -113,7 +113,7 @@
 
 #pragma mark - TableView Datasource
 
--(int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+-(NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
     return [arrRootCollection count];
 }
@@ -135,7 +135,7 @@
 
 #pragma mark - TableView Delegate
 
--(float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+-(CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     return 47;
 }
